whatsapp-birthday-server/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
httpx>=0.27.0,<1.0.0
pydantic-settings>=2.2.0,<3.0.0
slowapi>=0.1.9
